requirements.txt
# Requirements for RAG AI App
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.7
langchain>=0.1.11
langchain-google-genai>=0.0.9
langchain-community>=0.0.25
faiss-cpu>=1.8.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
tiktoken>=0.5.2
unstructured>=0.12.0
cachetools>=5.3.2
jinja2>=3.1.3
SQLAlchemy>=2.0.0
# Document processing requirements
pypdf>=4.0.1
docx2txt>=0.8
python-docx>=1.1.0
docx2pdf>=0.1.8
reportlab>=4.0.0
regex>=2023.10.3
nltk>=3.8.1
# Meta AI and embedding requirements
meta-ai-api>=0.1.0
sentence-transformers>=2.2.2